Use `u64` with `ObjectStorage` to support large SSTs (#485)
The next release of the `object_store` crate has breaking changes. It
changes the byte lengths from `usize` to `u64`. This patch updates
SlateDB to use the new `u64` types. I also added some paranoia with
`try_from` instead of `as u32`.

Depends on https://github.com/apache/arrow-rs/issues/6903

+++ src/manifest/store.rs
@@ -423,7 +423,7 @@ pub(crate) struct ManifestFileMetadata {
     pub(crate) location: Path,
     pub(crate) last_modified: chrono::DateTime<Utc>,
     #[allow(dead_code)]
-    pub(crate) size: usize,
+    pub(crate) size: u32,
 }
 
 fn serialize_path<S>(path: &Path, serializer: S) -> Result<S::Ok, S::Error>
@@ -511,7 +511,7 @@ impl ManifestStore {
                         id,
                         location: file.location,
                         last_modified: file.last_modified,
-                        size: file.size,
+                        size: file.size as u32,
                     });
                 }
                 Err(_) => warn!("Unknown file in manifest directory: {:?}", file.location),
